Extract shared ILInspector.ControlFlow (CFG edges + dominance kernel) (#1807)
Phase (a) of #1807: lift the IR-agnostic control-flow primitives — BlockEdges and
the Cooper-Harvey-Kennedy PostDominators kernel — out of ILInspector.Decompiler
into a new low-level ILInspector.ControlFlow library that depends on neither the
decompiler nor the analyzer. The decompiler keeps Cfg.Build(Block[]) producing
ControlFlow.BlockEdges and feeds the kernel; ILInspector.Analysis can do the same
from its own IL scan when phase (b) adds a data-flow shape.

Pure extraction, no behavior change: PostDominators tests (6) and CfgMermaid tests
pass unchanged; CLI + decompiler + harness build. The Block[] convenience moves to
call sites as Of(Cfg.Build(blocks)).
